Try: Fix an issue with ghost fields in Events table

+++ db/migrate/20251206001208_remove_events_field_from_events.rb
@@ -1,5 +1,11 @@
 class RemoveEventsFieldFromEvents < ActiveRecord::Migration[8.1]
   def change
-    remove_column :events, :events
+    if column_exists?(:events, :events)
+      remove_column :events, :events
+    end
+
+    if column_exists?(:events, :string)
+      remove_column :events, :string
+    end
   end
 end
